+++ crates/client/src/error/dnssec_error.rs
@@ -1,24 +1,13 @@
-/*
- * Copyright (C) 2015 Benjamin Fry <benjaminfry@me.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2015-2020 Benjamin Fry <benjaminfry@me.com>
+//
+// Licensed under the Apache License, Version 2.0, <LICENSE-APACHE or
+// http://apache.org/licenses/LICENSE-2.0> or the MIT license <LICENSE-MIT or
+// http://opensource.org/licenses/MIT>, at your option. This file may not be
+// copied, modified, or distributed except according to those terms.
 
 //! Dnssec error types for the crate
 
 use std::fmt;
-
-use failure::{Backtrace, Context, Fail};
 
 #[cfg(not(feature = "openssl"))]
 use self::not_openssl::SslErrorStack;
@@ -28,96 +17,95 @@ use self::not_ring::{KeyRejected, Unspecified};
 use openssl::error::ErrorStack as SslErrorStack;
 #[cfg(feature = "ring")]
 use ring::error::{KeyRejected, Unspecified};
+use thiserror::Error;
 use trust_dns_proto::error::{ProtoError, ProtoErrorKind};
+
+use crate::proto::{trace, ExtBacktrace};
 
 /// An alias for dnssec results returned by functions of this crate
 pub type Result<T> = ::std::result::Result<T, Error>;
 
 /// The error kind for dnssec errors that get returned in the crate
-#[derive(Eq, PartialEq, Debug, Fail)]
+#[derive(Debug, Error)]
 pub enum ErrorKind {
     /// An error with an arbitrary message, referenced as &'static str
-    #[fail(display = "{}", _0)]
+    #[error("{0}")]
     Message(&'static str),
 
     /// An error with an arbitrary message, stored as String
-    #[fail(display = "{}", _0)]
+    #[error("{0}")]
     Msg(String),
 
     // foreign
     /// An error got returned by the trust-dns-proto crate
-    #[fail(display = "proto error")]
-    Proto,
+    #[error("proto error: {0}")]
+    Proto(#[from] ProtoError),
 
     /// A ring error
-    #[fail(display = "ring error")]
-    Ring,
+    #[error("ring error: {0}")]
+    RingKeyRejected(#[from] KeyRejected),
+
+    /// A ring error
+    #[error("ring error: {0}")]
+    RingUnspecified(#[from] Unspecified),
 
     /// An ssl error
-    #[fail(display = "ssl error")]
-    SSL,
+    #[error("ssl error: {0}")]
+    SSL(#[from] SslErrorStack),
 
     /// A request timed out
-    #[fail(display = "request timed out")]
+    #[error("request timed out")]
     Timeout,
 }
 
 impl Clone for ErrorKind {
     fn clone(&self) -> Self {
         use self::ErrorKind::*;
-        match *self {
+        match self {
             Message(msg) => Message(msg),
             Msg(ref msg) => Msg(msg.clone()),
 
             // foreign
-            Proto => Proto,
-            Ring => Ring,
-            SSL => SSL,
+            Proto(proto) => Proto(proto.clone()),
+            RingKeyRejected(r) => Msg(format!("Ring rejected key: {}", r)),
+            RingUnspecified(_r) => RingUnspecified(Unspecified),
+            SSL(ssl) => Msg(format!("SSL had an error: {}", ssl)),
             Timeout => Timeout,
         }
     }
 }
 
 /// The error type for dnssec errors that get returned in the crate
-#[derive(Debug)]
+#[derive(Debug, Clone, Error)]
 pub struct Error {
-    inner: Context<ErrorKind>,
+    kind: ErrorKind,
+    backtrack: Option<ExtBacktrace>,
 }
 
 impl Error {
     /// Get the kind of the error
     pub fn kind(&self) -> &ErrorKind {
-        self.inner.get_context()
-    }
-}
-
-impl Fail for Error {
-    fn cause(&self) -> Option<&dyn Fail> {
-        self.inner.cause()
-    }
-
-    fn backtrace(&self) -> Option<&Backtrace> {
-        self.inner.backtrace()
+        &self.kind
     }
 }
 
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        fmt::Display::fmt(&self.inner, f)
+        if let Some(ref backtrace) = self.backtrack {
+            fmt::Display::fmt(&self.kind, f)?;
+            fmt::Debug::fmt(backtrace, f)
+        } else {
+            fmt::Display::fmt(&self.kind, f)
+        }
     }
 }
 
 impl From<ErrorKind> for Error {
     fn from(kind: ErrorKind) -> Error {
         Error {
-            inner: Context::new(kind),
+            kind,
+            backtrack: trace!(),
         }
-    }
-}
-
-impl From<Context<ErrorKind>> for Error {
-    fn from(inner: Context<ErrorKind>) -> Error {
-        Error { inner }
     }
 }
 
@@ -136,27 +124,27 @@ impl From<String> for Error {
 impl From<ProtoError> for Error {
     fn from(e: ProtoError) -> Error {
         match *e.kind() {
-            ProtoErrorKind::Timeout => e.context(ErrorKind::Timeout).into(),
-            _ => e.context(ErrorKind::Proto).into(),
+            ProtoErrorKind::Timeout => ErrorKind::Timeout.into(),
+            _ => ErrorKind::from(e).into(),
         }
     }
 }
 
 impl From<KeyRejected> for Error {
     fn from(e: KeyRejected) -> Error {
-        e.context(ErrorKind::Ring).into()
+        ErrorKind::from(e).into()
     }
 }
 
 impl From<Unspecified> for Error {
     fn from(e: Unspecified) -> Error {
-        e.context(ErrorKind::Ring).into()
+        ErrorKind::from(e).into()
     }
 }
 
 impl From<SslErrorStack> for Error {
     fn from(e: SslErrorStack) -> Error {
-        e.context(ErrorKind::SSL).into()
+        ErrorKind::from(e).into()
     }
 }